Move implicits and test results out of MultiVariantTests
