+++ app/com/gu/identity/frontend/mvt/MultiVariantTestAction.scala
@@ -13,7 +13,7 @@ object MultiVariantTestRequest {
 
   def apply[A](request: Request[A]): MultiVariantTestRequest[A] = {
     val mvtCookie = getMvtCookie(request)
-    val activeTests = mvtCookie.map(id => MultiVariantTests.activeTests(id)).getOrElse(Nil).toMap
+    val activeTests = mvtCookie.map(id => TestResults.activeTests(id)).getOrElse(Nil).toMap
 
     MultiVariantTestRequest[A](mvtCookie, activeTests ++ getTestOverrides(request.queryString), request)
   }
